+++ index.php
@@ -78,6 +78,7 @@
                       <ul class="list-group">
                           <li class="list-group-item">Classic regular example<br>(<a href="scenarios/regular/GET/classic.php">GET</a> | <a href="scenarios/regular/POST/classic.php"> POST</a>)</li>
                           <li class="list-group-item">Classic (Base64) regular example<br>(<a href="scenarios/regular/GET/classic_b64.php">GET</a> | <a href="scenarios/regular/POST/classic_b64.php"> POST</a>)</li>
+                          <li class="list-group-item">Classic (Hex) regular example<br>(<a href="scenarios/regular/GET/classic_hex.php">GET</a> | <a href="scenarios/regular/POST/classic_hex.php"> POST</a>)</li>
                           <li class="list-group-item">Classic single-quote example<br>(<a href="scenarios/regular/GET/classic_quote.php">GET</a> | <a href="scenarios/regular/POST/classic_quote.php"> POST</a>)</li>       
                           <li class="list-group-item">Classic double-quote example<br>(<a href="scenarios/regular/GET/classic_double_quote.php">GET</a> | <a href="scenarios/regular/POST/classic_double_quote.php"> POST</a>)</li> 
                           <li class="list-group-item">Classic non-space example<br>(<a href="scenarios/regular/GET/classic_non_space.php">GET</a> | <a href="scenarios/regular/POST/classic_non_space.php"> POST</a>)</li>